// src/Enums/ActionsPropType.ts
export enum ActionType {
		ExecFunc = 'exec-function',
		Flow = 'action-flow',
		Mass = 'mass',
		RunScript = 'run-script',
}


export enum ActionPropType {
		Name = 'name',
		Type = 'type',
		Comment = 'comment',
		Model = 'model',
		Lib = 'lib',
		ctx = 'ctx',
		Func = 'func',
		Farg = 'farg',
		Output = 'output',
		Steps = 'steps',
}